+++ src/main/java/cn/mybase/ssm/service/impl/dao/impl/mybatis/PermissionMapper.java
@@ -14,7 +14,7 @@ public interface PermissionMapper {
 
     int deleteByExample(PermissionExample example);
 
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(Permission record);
 
@@ -22,7 +22,7 @@ public interface PermissionMapper {
 
     List<Permission> selectByExample(PermissionExample example);
 
-    Permission selectByPrimaryKey(Long id);
+    Permission selectByPrimaryKey(Integer id);
 
     int updateByExampleSelective(@Param("record") Permission record, @Param("example") PermissionExample example);
